requirements.txt
asgiref==3.7.2
certifi==2023.11.17 
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-json-widget==1.1.1
django-multiselectfield==0.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
future==0.18.3
gunicorn==21.2.0
idna==3.6
packaging==23.2
Pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
strawberry-graphql[django]==0.273.2
graphql-core==3.2.3
typing_extensions==4.12.2
urllib3==2.1.0
whitenoise==6.6.0
django-unfold==0.47.0
bw_simapro_csv==0.3.7
psutil==7.0.0
python-decouple==3.8